+++ Changelog.md
@@ -22,6 +22,7 @@ Changelog [format](http://keepachangelog.com/en/1.0.0/)
 * GODT-363 Drafts are not deleted when already created on webapp.
 * GODT-390 Don't logout user if AuthRefresh fails because internet was off
 * GODT-341 Fixed flaky unittest for Store synchronization cooldown
+* Crash when failing to match necessary html element
 
 ## [v1.2.7] Donghai-hotfix - beta (2020-05-07)